+++ CHANGES.md
@@ -1,5 +1,21 @@
 # Changes history
 
+### 2.2.0
+
+* Added `KHR_materials_transmission` extension support.
+* Added `KHR_materials_volume` extension support.
+* Added `KHR_materials_ior` extension support.
+* Added `KHR_materials_specular` extension support.
+* Added `KHR_materials_iridescence` extension support.
+* Added `KHR_materials_emissive_strength` extension support.
+* Added optional TransmissionSource renderer to SceneManager for scene refraction effect.
+* Added shader parser in order to split PBR shader into several files for better maintainability.
+* Added HDR/RGBE subformat support
+* Added USHORT and UBYTE support for bone weights
+* Fix PBR shader: Ambient occlusion texture is now applied to IBL only instead of all lights.
+* Added Mirror specular effect (dynamic reflections).
+* Added cascade shadow map.
+
 ### 2.1.0
 
 * Added optional environement and skybox rotation.